CS-2243-speed-up-product-group-sync: - Process product groups in batches - Refactored existing API utility functions to improve modularity and maintainability.
